Moving group modal components to separate folder
