+++ upnext-web-app/src/UpNext/components/GroupModals/DeleteGroupModal.tsx
