+++ upnext-web-app/src/UpNext/components/styles/GroupDetailsModal.css
